Handle case where time is "--"
When PDX says a time is "--", it sets the Timeslot's calendar object to
null. When a Timeslot's calendar object is null, it displays on the main
activity as "--".

We'll probably have to change extractPDXMetalsContent to be a boolean to
see whether or not there's a "--" in the times, so we know whether to
try to set alarms again later.