Fixed cron job, hopefully
When alarm goes off, app tries to pull PDX homepage and parse it for
relevant metals/godfest info.

Alarm is set to go off once a day between 12:30 am and 1:30 am every
day, at a fixed time stored in shared preferences. (This doesn't change
until the app is reinstalled.) When the alarm is set, it checks to see
if that time has already passed--if so, it sets the alarm to go off
tomorrow at that time.